requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.1
requests==2.32.3
Werkzeug==3.0.3
Flask-Caching==2.3.0
python-dotenv==1.0.1
gunicorn==22.0.0
